mv build_index to ann_util.py
